add addTailAnimationDelay

+++ src/util/TargetTailAnimation.java
@@ -12,7 +12,7 @@ public class TargetTailAnimation implements IRenderable,Destroyable {
 	private int y;
 	
 	private int min_radius = Configuration.TAIL_MIN_RADIUS;
-	private int radius = Configuration.TAIL_MAX_RADIUS;
+	private double radius = Configuration.TAIL_MAX_RADIUS;
 	
 	public TargetTailAnimation(int x,int y){
 		this.x = x;
@@ -29,7 +29,7 @@ public class TargetTailAnimation implements IRenderable,Destroyable {
 	}
 	@Override
 	public void draw(Graphics2D g2) {
-		DrawingUtility.drawCircle(g2, x, y, 0, radius, Color.WHITE, Color.ORANGE);
+		DrawingUtility.drawCircle(g2, x, y, 0, (int)radius, Color.WHITE, Color.ORANGE);
 		radius -= Configuration.TAIL_RADIUS_REDUCE_RATE;
 	}
